Revert "Создаем модули'"
This reverts commit 863e3f4dc80d809ebf3e91ef0b0e4065e2e567e9.
